+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.9] - 2020-03-11
+
+### Added
+- `padding` prop // padding is applied to the inner SlideOut container
+- `overlayColor` prop // change default color to overlay takes in HEX color code (example: "#ffffff")
+- `overlayOpacity` prop // change defauly opacity to overlay
+
+### Fixed
+- `noOverlay`
+
 ## [0.0.8] - 2020-03-10
 
 ### Added